+++ addon/components/bootstrap-inputs/-date.js
@@ -1,5 +1,6 @@
 import Component from '@ember/component';
 import layout from '../../templates/components/bootstrap-inputs/-date';
+import { or } from '@ember/object/computed';
 import { PropTypes } from 'ember-prop-types';
 import {
   BuilderForPropTypes,
@@ -83,6 +84,7 @@ export const propDefinitions = {
 export default Component.extend({
   layout,
   propTypes: BuilderForPropTypes(propDefinitions),
+  _disabled: or('formDisabled', 'disabled'),
 
   todayHighlight: true,
   format: 'mm/dd/yyyy',