+++ addon/components/bootstrap-inputs/-tel.js
@@ -1,5 +1,6 @@
 import Component from '@ember/component';
 import layout from '../../templates/components/bootstrap-inputs/-tel';
+import { or } from '@ember/object/computed';
 import { PropTypes } from 'ember-prop-types';
 import {
   BuilderForPropTypes,
@@ -72,6 +73,7 @@ export const propDefinitions = {
 export default Component.extend({
   layout,
   propTypes: BuilderForPropTypes(propDefinitions),
+  _disabled: or('formDisabled', 'disabled'),
 
   getDefaultProps() {
     return BuilderForPropDefaults(propDefinitions);